+++ src/Traits/HasSettings.php
@@ -2,11 +2,25 @@
 
 namespace Glorand\Model\Settings\Traits;
 
-use Glorand\Model\Settings\Managers\AbstractSettingsManager;
+use Glorand\Model\Settings\Contracts\SettingsManagerContract;
 use Illuminate\Support\Arr;
 
+/**
+ * @property array $settingsRules
+ * @property array $defaultSettings
+ * @SuppressWarnings(PHPMD.StaticAccess)
+ */
 trait HasSettings
 {
+    public function getRules(): array
+    {
+        if (property_exists($this, 'settingsRules') && is_array($this->settingsRules)) {
+            return $this->settingsRules;
+        }
+
+        return [];
+    }
+
     public function getDefaultSettings(): array
     {
         if (property_exists($this, 'defaultSettings')
@@ -16,10 +30,20 @@ trait HasSettings
             && is_array($defaultSettings)) {
             return Arr::wrap($defaultSettings);
         }
+
         return [];
+    }
+
+    public function __call($name, $args)
+    {
+        if (isset($this->invokeSettingsBy) && $name === $this->invokeSettingsBy) {
+            return $this->settings();
+        }
+
+        return call_user_func(get_parent_class($this) . '::__call', $name, $args);
     }
 
     abstract public function getSettingsValue(): array;
 
-    abstract public function settings(): AbstractSettingsManager;
+    abstract public function settings(): SettingsManagerContract;
 }